+++ comprehensive-review/solutions/garden-back/src/main/java/com/redhat/garden/back/event/LowTemperatureDetected.java
@@ -4,11 +4,11 @@ public class LowTemperatureDetected {
     public String gardenName;
     public Integer sensorId;
     public Double value;
-    public Integer timestamp;
+    public Long timestamp;
 
     public LowTemperatureDetected() {}
 
-    public LowTemperatureDetected(String gardenName, Integer sensorId, Double value,  Integer timestamp) {
+    public LowTemperatureDetected(String gardenName, Integer sensorId, Double value,  Long timestamp) {
         this.gardenName = gardenName;
         this.sensorId = sensorId;
         this.value = value;
