+++ comprehensive-review/solutions/garden-back/src/main/java/com/redhat/garden/back/event/front/GardenEvent.java
@@ -4,11 +4,11 @@ public class GardenEvent {
     public String name;
     public String garden;
     public Integer sensorId;
-    public Integer timestamp;
+    public Long timestamp;
 
     public GardenEvent() {}
 
-    public GardenEvent(String name, String garden, Integer sensorId, Integer timestamp) {
+    public GardenEvent(String name, String garden, Integer sensorId, Long timestamp) {
         this.name = name;
         this.garden = garden;
         this.sensorId = sensorId;
