+++ comprehensive-review/solutions/garden-back/src/main/java/com/redhat/garden/back/event/front/GardenStatus.java
@@ -6,11 +6,11 @@ public class GardenStatus {
     public Double temperature;
     public String garden;
     public Integer sensorId;
-    public Integer timestamp;
+    public Long timestamp;
 
     public GardenStatus() {}
 
-    public GardenStatus(Integer id, String name, Double temperature, String garden, Integer sensorId, Integer timestamp) {
+    public GardenStatus(Integer id, String name, Double temperature, String garden, Integer sensorId, Long timestamp) {
         this.id = id;
         this.name = name;
         this.temperature = temperature;
