+++ comprehensive-review/solutions/garden-back/src/main/java/com/redhat/garden/back/measurement/SensorMeasurement.java
@@ -4,11 +4,11 @@ public class SensorMeasurement {
     public Integer sensorId;
     public String property;
     public Double value;
-    public Integer timestamp;
+    public Long timestamp;
 
     public SensorMeasurement() {}
 
-    public SensorMeasurement(Integer sensorId, String property, Double value, Integer timestamp) {
+    public SensorMeasurement(Integer sensorId, String property, Double value, Long timestamp) {
         this.sensorId = sensorId;
         this.property = property;
         this.value = value;
